requirements.txt
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
typing-extensions>=4.8.0
ormsgpack>=0.4.0

